Move ts tests to test/ts/
